refactored existing code
