+++ src/db/mod.rs
@@ -1,3 +1,6 @@
+pub mod schema;
+pub mod sql_types;
+
 use diesel::pg::PgConnection;
 
 #[database("postgres")]
