+++ src/db/schema.rs
@@ -1,6 +1,7 @@
 table! {
+
     use diesel::sql_types::*;
-    use crate::sql_types::*;
+    //use crate::db::sql_types::*;
 
     events (id) {
         id -> Int4,
@@ -13,8 +14,9 @@ table! {
 }
 
 table! {
+
     use diesel::sql_types::*;
-    use crate::sql_types::*;
+    use crate::db::sql_types::*;
 
     permissions (id) {
         id -> Int4,
@@ -26,8 +28,9 @@ table! {
 }
 
 table! {
+
     use diesel::sql_types::*;
-    use crate::sql_types::*;
+    //use crate::db::sql_types::*;
 
     projects (id) {
         id -> Int4,
@@ -40,8 +43,9 @@ table! {
 }
 
 table! {
+
     use diesel::sql_types::*;
-    use crate::sql_types::*;
+    //use crate::db::sql_types::*;
 
     roles (id) {
         id -> Int4,
@@ -50,8 +54,9 @@ table! {
 }
 
 table! {
+
     use diesel::sql_types::*;
-    use crate::sql_types::*;
+    ////use crate::db::sql_types::*;
 
     submissions (id) {
         id -> Int4,
@@ -61,8 +66,9 @@ table! {
 }
 
 table! {
+
     use diesel::sql_types::*;
-    use crate::sql_types::*;
+    //use crate::db::sql_types::*;
 
     user_events (id) {
         id -> Int4,
@@ -72,8 +78,9 @@ table! {
 }
 
 table! {
+
     use diesel::sql_types::*;
-    use crate::sql_types::*;
+    //use crate::db::sql_types::*;
 
     user_roles (id) {
         id -> Int4,
@@ -83,8 +90,9 @@ table! {
 }
 
 table! {
+
     use diesel::sql_types::*;
-    use crate::sql_types::*;
+    //use crate::db::sql_types::*;
 
     users (id) {
         id -> Int4,
