+++ src/main.rs
@@ -33,8 +33,8 @@ mod github;
 mod resolvers;
 mod routes;
 #[allow(unused_imports)]
-mod schema;
-mod sql_types;
+// mod schema;
+// mod sql_types;
 mod types;
 mod utils;
