+++ src/types/submission.rs
@@ -1,4 +1,4 @@
-use crate::schema::*;
+use crate::db::schema::*;
 use crate::types::{Project, RawUser, User};
 use diesel::{self, AsChangeset, Queryable};
 use serde_derive::{Deserialize, Serialize};
